reward: Relocate variables
